Merge flexget installer/updater changes

+++ Flexget/flexget-install.sh
@@ -1,7 +1,7 @@
 #!/bin/sh
 
-# Flexget Installer by Xan#7777
-# This installs flexget under python-venv
+# Flexget Installer/Updater by Xan#7777
+# This installs/updates flexget under python-venv
 
 # Unofficial Script warning
 clear
@@ -30,18 +30,48 @@ fi
 clear
 echo "Installing flexget..."
 sleep 1
-python -m venv "$HOME"/flexget/
-cd "$HOME"/flexget || exit
+if [ ! -d "$HOME"/flexget/ ]; then
+    python -m venv "$HOME"/flexget/
+fi
 "$HOME"/flexget/bin/pip list --outdated --format=freeze | grep -v '^\-e' | cut -d = -f 1 | xargs -n1 "$HOME"/flexget/bin/pip install -U 
-"$HOME"/flexget/bin/pip install wheel
+"$HOME"/flexget/bin/pip install wheel --upgrade
 "$HOME"/flexget/bin/pip install flexget --upgrade
-ln -s "$HOME"/flexget/bin/flexget "$HOME"/bin/flexget
+if [ ! -f "$HOME"/bin/flexget ]; then
+    ln -s "$HOME"/flexget/bin/flexget "$HOME"/bin/flexget
+fi
 
 # Check flexget
 clear
 command -v flexget
 flexget --version
-sleep 1
+sleep 2
+
+# Client Installer
+clear
+echo "You may need to install a client that can interface with your actual torent client."
+echo "Choose between deluge or transmission. You may also skip if you do not need it."
+
+while true; do
+read -p "Enter your response here: " flexent
+    case $flexent in
+        deluge)
+            "$HOME"/flexget/bin/python -m pip install deluge-client --upgrade
+            echo "Refer to this page for more info: https://flexget.com/Plugins/deluge"
+            sleep 7
+            break
+            ;;
+        transmission)
+            "$HOME"/flexget/bin/python -m pip install transmissionrpc --upgrade
+            echo "Refer to this page for more info: https://flexget.com/Plugins/transmission"
+            sleep 7
+            break
+            ;;
+        skip)
+            break
+            ;;
+        *) echo "Unknown response. Try again..." ;;
+    esac
+done
 
 # Cleanup and exit
 clear
